backend/requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
certifi==2020.12.5
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.11.1
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
promise==2.3
psycopg2-binary==2.8.5
PyJWT[crypto]==1.7.1
pytz==2020.1
requests==2.24.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.11